Moved BDM to hdiv
